+++ packages/components/molecules/f-mega-modal/stories/MegaModal.stories.js
@@ -24,7 +24,7 @@ export const MegaModalComponent = (args, { argTypes }) => ({
             :has-overlay="hasOverlay"
             :has-close-button="hasCloseButton"
             :has-back-button="hasBackButton"
-            :has-mode-right-to-left="hasModeRightToLeft"
+            :is-mode-right-to-left="isModeRightToLeft"
             :close-on-blur="closeOnBlur"
             :close-button-copy="closeButtonCopy"
             :title="titleCopy"
@@ -52,7 +52,7 @@ MegaModalComponent.args = {
     hasOverlay: true,
     hasCloseButton: true,
     hasBackButton: false,
-    hasModeRightToLeft: false,
+    isModeRightToLeft: false,
     closeOnBlur: true,
     closeButtonCopy: 'Close modal',
     titleCopy: 'This place isn’t taking orders',
@@ -96,7 +96,7 @@ MegaModalComponent.argTypes = {
     hasBackButton: {
         control: { type: 'boolean' }
     },
-    hasModeRightToLeft: {
+    isModeRightToLeft: {
         control: { type: 'boolean' }
     },
     closeOnBlur: {